refactor: ShortestPath 삭제
