+++ README.md
@@ -54,8 +54,8 @@
     - [x] 노선에 A-B-C 역이 연결되어 있을 때 B역을 제거할 경우 A-C로 재배치 된다.
     - [x] 거리는 두 구간의 거리의 합으로 정한다.
 
-  
-### 최단 경로 탐색기 (PathFinder)
+
+### 최단 경로 탐색기 (ShortestPathFinder)
 - [x] 출발역과 도착역이 같은 경우 경로를 구할 수 없다.
 - [x] 출발역과 도착역이 연결되어 있지 않은 경우 경로를 구할 수 없다.
 - [x] 출발역과 도착역을 연결하는 최단 경로를 구한다.
